Page shows users list and validates if the token is stored

+++ src/main.jsx
@@ -16,15 +16,19 @@ import './assets/styles.css';
 //*Method to verify token
 const verifySession = ()=>{
   const token = sessionStorage.getItem('token');
-  const decoded = jwtDecode(token);
-  
-  const currentDate = Math.floor(Date.now() / 1000);
-  //*If token is no longer valid, it'll return false
-  if(decoded.exp < currentDate){
-    sessionStorage.clear();
-    return false;
+  //*In case the token exists
+  if(token){
+    const decoded = jwtDecode(token);
+
+    const currentDate = Math.floor(Date.now() / 1000);
+    //*If token is no longer valid, it'll return false
+    if(decoded.exp < currentDate){
+      sessionStorage.clear();
+      return false;
+    }
+    return true
   }
-  return true
+  return false;
 }
 
 //*Create router
